requirements.txt
argon2-cffi==18.3.0
asn1crypto==0.24.0
astroid==2.2.5
asttokens==1.1.15
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.11.5
colorama==0.4.1
cryptography==2.4.2
cx-Freeze==6.0
icontract==2.0.6
idna==2.7
isort==4.3.15
Jinja2==2.10.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
mypy-extensions==0.4.3
passlib==1.7.1
pep8==1.7.1
pipdeptree==0.13.2
pluggy==0.9.0
py==1.8.0
pycodestyle==2.4.0
pycparser==2.19
pygame==1.9.4
pygame-menu==2.1.0
pylint==2.3.1
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.4
pyOpenSSL==18.0.0
pyperclip==1.7.0
pytest==4.3.1
PyYAML==5.1.2
six==1.11.0
swagger-to==3.1.2
typed-ast==1.3.1
typing-extensions==3.7.4.1
typing-inspect==0.5.0
wrapt==1.11.1
